+++ CHANGES.rst
@@ -1,13 +1,16 @@
 Release Notes
 =============
 
+3.3.0
+-----
+- Added Cover Element keyword to hide elements without breaking page layout [SergiuTudos]
+
 3.2.0
 -----
 - Added message param to keywords `Location Should Be` and `Location Should Contain` to display custom error message [taniabhullar]
 - Added `Element Attribute Value Should Be` verifies element identified by locator contains expected attribute value.. [brian-mathews]
 - Added `Get Session Id` keyword to get remote webdriver session id [ilfirinpl]
 - Fixed example documentation for custom locator [DanielPBak]
-- Added Cover Element keyword to hide elements without breaking page layout [SergiuTudos]
 
 3.1.1
 -----